All tests have been merged into test_example_selector.py
